move factory to dkan_common. and make a dkan_common.
