+++ modules/custom/dkan_api/dkan_api.info.yml
@@ -5,4 +5,5 @@ core: 8.x
 dependencies:
   - dkan_data
   - basic_auth
+  - dkan_common
 package: DKAN
