+++ modules/custom/dkan_common/dkan_common.info.yml
@@ -1,0 +1,4 @@
+name: DKAN Common
+description: 'Common dkan utilities and function.'
+type: module
+core: 8.x
