+++ modules/custom/dkan_common/dkan_common.services.yml
@@ -1,0 +1,4 @@
+services:
+    dkan.factory:
+        class: Drupal\dkan_common\Service\DkanFactory
+        arguments: ['@service_container']
