+++ modules/custom/dkan_datastore/tests/src/Unit/SqlParserTest.php
@@ -9,8 +9,6 @@ use PHPUnit\Framework\TestCase;
  */
 class SqlParserTest extends TestCase {
 
-  use \Dkan\PhpUnit\DkanUnitTestTrait;
-
   /**
    *
    */
